requirements.txt
Django==3.0.4
DateTime==4.3
python-dateutil==1.4
